Introduce an Interface for  RegistrationEngine
